fix: mips和sw_64环境下不让连接sanitize库
mips和sw_64环境下取消对sanitize库的连接，通过宏来指定是否生成内存检测文件

Log: mips和sw_64环境下不让连接sanitize库

+++ tests/test-recoverage.sh
@@ -25,5 +25,5 @@ cd ..
 genhtml -o $REPORT_DIR $BUILD_DIR/coverage.info
 #rm -rf $BUILD_DIR
 #rm -rf ../$BUILD_DIR
-
-mv ./$BUILD_DIR/asan.log* ./$BUILD_DIR/asan_dde-network-utils.log
+[ -e ./$BUILD_DIR/asan.log* ] && mv ./$BUILD_DIR/asan.log* ./$BUILD_DIR/asan_dde-network-utils.log || touch ./$BUILD_DIR/asan_dde-network-utils.log
+#mv ./$BUILD_DIR/asan.log* ./$BUILD_DIR/asan_dde-network-utils.log
